KATA-1147: PID-related UI-redesign
Access PID moved away from pids and renamed to external_id. This is because we cannot know what the future brings for this identifier. From the add dataset viewpoint, this id is not input where other identifiers are given. It is currently a hidden input, whose value is changed currently only when reetta ida availability related pid is given. It may be in the future if other types of identifiers need to be added as external_id, additional logic to populate the hidden field is required. External id is now shown in show dataset view only if the user has rights to update the package.